fix(frontend): add JSDoc documentation to wallet connect button props (#43)
